\name{edcc-class}
\docType{class}
\alias{edcc-class}
\title{Class "edcc"}
\description{A list of objects of class \code{edcc}.}
\section{Objects from the Class}{
  Objects can be created by calling the \code{ecoXbar}, \code{ecoCusum}
  or \code{ecoEwma} function.
}
\section{Methods}{
  \describe{
    \item{\link[graphics]{contour}}{: creat a contour plot.}
    \item{\link[stats]{update}}{: update the object.}
    \item{\link[base]{print}}{: print the object.}
  }
}
\keyword{classes}
